Organize package-wide WMO tests
