+++ pyproject.toml
@@ -62,7 +62,8 @@ include = [
 ]
 
 [tool.pytest.ini_options]
-# Tests live inline next to the modules they cover (foo.py -> foo_test.py).
+# Tests live inline next to the modules they cover (foo.py -> foo_test.py), or in a package's tests
+# directory when they cover the package as a whole.
 python_files = ["*_test.py"]
 # One root gate over the single package.
 testpaths = ["wmo"]
