+++ wmo/cli/tests/startup_test.py
